+++ src/imp/mod.rs
@@ -112,6 +112,10 @@ mod fallback;
 )]
 #[cfg(any(
     portable_atomic_armv6m,
+    all(
+        target_arch = "arm",
+        not(any(target_feature = "v6", portable_atomic_target_feature = "v6"))
+    ),
     target_arch = "avr",
     target_arch = "msp430",
     target_arch = "riscv32",