+++ tools/build.sh
@@ -166,7 +166,7 @@ build() {
                 RUSTFLAGS="${target_rustflags}" \
                     x cargo "${args[@]}" --manifest-path tests/api-test/Cargo.toml "$@"
                 ;;
-            bpf* | thumbv4t-*) ;; # TODO
+            bpf*) ;; # TODO
             *)
                 RUSTFLAGS="${target_rustflags} --cfg portable_atomic_unsafe_assume_single_core" \
                     x cargo "${args[@]}" --manifest-path tests/api-test/Cargo.toml "$@"
@@ -196,7 +196,7 @@ build() {
             if ! grep <<<"${cfgs}" -q "target_has_atomic=" && [[ -n "${has_asm}" ]]; then
                 case "${target}" in
                     avr-* | msp430-*) ;;  # always single-core
-                    bpf* | thumbv4t-*) ;; # TODO
+                    bpf*) ;; # TODO
                     *)
                         RUSTFLAGS="${target_rustflags} --cfg portable_atomic_unsafe_assume_single_core" \
                             x cargo "${args[@]}" --target-dir target/assume-single-core "$@"